requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
coverage==7.13.2
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
